requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
distlib==0.3.7
filelock==3.12.4
idna==3.4
lxml==4.9.3
meson==1.2.2
Pillow==10.1.0
platformdirs==3.11.0
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pyTelegramBotAPI==4.14.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
urllib3==2.0.7
virtualenv==20.24.5
